feat: admin dashboard

+++ src/pages/dashboard/Manager/IndexTicket.js
@@ -24,7 +24,14 @@ import {
   Delete,
   ViewCompact,
 } from "@mui/icons-material";
-import { Box, FormControl, MenuItem, Pagination, Select } from "@mui/material";
+import {
+  Box,
+  FormControl,
+  InputLabel,
+  MenuItem,
+  Pagination,
+  Select,
+} from "@mui/material";
 import { toast } from "react-toastify";
 import { useNavigate } from "react-router-dom";
 import PageSizeSelector from "../Pagination/Pagination";
@@ -36,6 +43,7 @@ import {
 import CustomizedProgressBars from "../../../components/iconify/LinearProccessing";
 import { useSelector } from "react-redux";
 import { formatDate } from "../../helpers/FormatDate";
+import CircularLoading from "../../../components/iconify/CircularLoading";
 
 const IndexTicket = () => {
   const [dataTickets, setDataTickets] = useState([]);
@@ -47,7 +55,7 @@ const IndexTicket = () => {
   const [searchField, setSearchField] = useState("title");
   const [searchQuery, setSearchQuery] = useState("");
   const [sortDirection, setSortDirection] = useState("desc");
-  const [sortBy, setSortBy] = useState("id");
+  const [sortBy, setSortBy] = useState("createdAt");
   const [ticketStatus, setTicketStatus] = useState(null);
   const [selectedTickets, setSelectedTickets] = useState([]);
   const [selectedTicketId, setSelectedTicketId] = useState(null);
@@ -72,7 +80,8 @@ const IndexTicket = () => {
         sortDirection,
         ticketStatus
       );
-      setDataTickets(res);
+      setDataTickets(res?.data);
+      setTotalPages(res?.totalPage);
       setIsLoading(false);
     } catch (error) {
       console.log("Error while fetching data", error);
@@ -87,7 +96,7 @@ const IndexTicket = () => {
   const fetchAllCategories = async () => {
     try {
       const res = await getAllCategories();
-      setDataCategories(res);
+      setDataCategories(res?.data);
     } catch (error) {
       console.log("Error while fetching data", error);
     }
@@ -112,35 +121,42 @@ const IndexTicket = () => {
   }, [selectedTickets]);
 
   const handleDeleteSelectedTickets = async (id) => {
-    try {
-      if (selectedTickets.length === 0) {
-        return;
-      }
-      const deletePromises = selectedTickets.map(async (ticketId) => {
-        try {
-          const res = await deleteTicketByManager(ticketId);
-          if (res.isError) {
-            toast.error(
-              `Error deleting ticket with ID ${ticketId}: `,
-              res.message
-            );
-          }
-          toast.success(`Delete ticket with ID successful ${ticketId}`);
-          return ticketId;
-        } catch (error) {
-          toast.error(`Error deleting ticket with ID ${ticketId}: `, error);
-          return null;
+    const shouldDelete = window.confirm(
+      "Are you sure want to delete selected tickets"
+    );
+    if (shouldDelete) {
+      try {
+        if (selectedTickets.length === 0) {
+          return;
         }
-      });
+        const deletePromises = selectedTickets.map(async (ticketId) => {
+          try {
+            const res = await deleteTicketByManager(ticketId);
+            if (res.isError) {
+              toast.error(
+                `Error deleting ticket with ID ${ticketId}: `,
+                res.message
+              );
+            }
+            toast.success(`Delete ticket with ID successful ${ticketId}`);
+            return ticketId;
+          } catch (error) {
+            toast.error(`Error deleting ticket with ID ${ticketId}: `, error);
+            return null;
+          }
+        });
 
-      const deletedTicketIds = await Promise.all(deletePromises);
-      const updatedTickets = dataTickets.filter(
-        (ticket) => !deletedTicketIds.includes(ticket.id)
-      );
-      setDataTickets(updatedTickets);
-      setSelectedTickets([]);
-    } catch (error) {
-      toast.error("Failed to delete selected tickets, Please try again later");
+        const deletedTicketIds = await Promise.all(deletePromises);
+        const updatedTickets = dataTickets.filter(
+          (ticket) => !deletedTicketIds.includes(ticket.id)
+        );
+        setDataTickets(updatedTickets);
+        setSelectedTickets([]);
+      } catch (error) {
+        toast.error(
+          "Failed to delete selected tickets, Please try again later"
+        );
+      }
     }
   };
 
@@ -198,7 +214,6 @@ const IndexTicket = () => {
   useEffect(() => {
     fetchAllTicket();
     fetchAllCategories();
-    setTotalPages(4);
   }, [fetchAllTicket]);
 
   return (
@@ -208,33 +223,62 @@ const IndexTicket = () => {
           <MDBContainer fluid>
             <MDBNavbarBrand style={{ fontWeight: "bold", fontSize: "24px" }}>
               <ContentCopy style={{ marginRight: "20px", color: "#FFFFFF" }} />{" "}
-              <span style={{ color: "#FFFFFF" }}>All My Request</span>
+              <span style={{ color: "#FFFFFF" }}>All Tickets</span>
             </MDBNavbarBrand>
-            <MDBNavbarNav className="ms-auto manager-navbar-nav">
+            <MDBNavbarNav className="ms-auto manager-navbar-nav justify-content-end align-items-center">
               <MDBBtn
                 color="#eee"
                 style={{
                   fontWeight: "bold",
-                  fontSize: "20px",
+                  fontSize: 16,
                   color: "#FFFFFF",
                 }}
                 onClick={handleOpenRequestTicket}
               >
                 <FaPlus /> New
               </MDBBtn>
-              {isUserRole2 && (
-                <MDBBtn
-                  color="eee"
-                  style={{
-                    fontWeight: "bold",
-                    fontSize: "20px",
-                    color: "#FFFFFF",
-                  }}
-                  onClick={handleDeleteSelectedTickets}
+
+              <MDBBtn
+                color="eee"
+                style={{
+                  fontWeight: "bold",
+                  fontSize: 16,
+                  color: "#FFFFFF",
+                }}
+                onClick={handleDeleteSelectedTickets}
+              >
+                <Delete /> Delete
+              </MDBBtn>
+
+              <div style={{ textAlign: "center" }}>
+                <FormControl
+                  variant="outlined"
+                  size="small"
+                  sx={{ color: "white", margin: 1, width: "10vw" }}
                 >
-                  <Delete /> Delete
-                </MDBBtn>
-              )}
+                  <InputLabel sx={{ color: "white" }}>Select Status</InputLabel>
+                  <Select
+                    label={"Select Status"}
+                    value={ticketStatus}
+                    onChange={handleStatusChange}
+                    inputProps={{
+                      name: "sortField",
+                      id: "search-field",
+                    }}
+                    sx={{
+                      color: "white",
+                    }}
+                  >
+                    <MenuItem value={null}>All</MenuItem>
+                    <MenuItem value={0}>Open</MenuItem>
+                    <MenuItem value={1}>Assign</MenuItem>
+                    <MenuItem value={2}>Progress</MenuItem>
+                    <MenuItem value={3}>Resolved</MenuItem>
+                    <MenuItem value={4}>Closed</MenuItem>
+                    <MenuItem value={5}>Cancelled</MenuItem>
+                  </Select>
+                </FormControl>
+              </div>
               <FormControl
                 variant="outlined"
                 style={{
@@ -281,50 +325,17 @@ const IndexTicket = () => {
                 pageSize={pageSize}
                 handleChangePageSize={handleChangePageSize}
               />
-              <div style={{ textAlign: "center" }}>
-                <label
-                  style={{
-                    fontWeight: "bold",
-                    marginTop: "15px",
-                    color: "white",
-                  }}
-                >
-                  Sort Status Ticket:
-                </label>
-                <FormControl
-                  variant="outlined"
-                  style={{
-                    minWidth: 120,
-                    marginRight: 10,
-                    marginTop: 10,
-                    marginLeft: 10,
-                  }}
-                  size="small"
-                >
-                  <Select
-                    value={ticketStatus}
-                    onChange={handleStatusChange}
-                    inputProps={{
-                      name: "sortField",
-                      id: "search-field",
-                    }}
-                    style={{ color: "white" }}
-                  >
-                    <MenuItem value={null}>-</MenuItem>
-                    <MenuItem value={0}>Open</MenuItem>
-                    <MenuItem value={1}>Assign</MenuItem>
-                    <MenuItem value={2}>Progress</MenuItem>
-                    <MenuItem value={3}>Resolved</MenuItem>
-                    <MenuItem value={4}>Closed</MenuItem>
-                    <MenuItem value={5}>Cancelled</MenuItem>
-                  </Select>
-                </FormControl>
-              </div>
             </MDBNavbarNav>
           </MDBContainer>
         </MDBNavbar>
         {isLoading ? (
-          <CustomizedProgressBars />
+          <MDBTableBody className="bg-light">
+            <tr>
+              <td>
+                <CircularLoading />
+              </td>
+            </tr>
+          </MDBTableBody>
         ) : (
           <MDBTable
             className="align-middle mb-0"
@@ -340,8 +351,8 @@ const IndexTicket = () => {
                     onChange={handleSelectAllTickets}
                   />
                 </th>
-                <th style={{ fontWeight: "bold" }}></th>
-                <th
+
+                {/* <th
                   style={{ fontWeight: "bold" }}
                   className="sortable-header"
                   onClick={() => handleSortChange("id")}
@@ -354,7 +365,7 @@ const IndexTicket = () => {
                     ) : (
                       <ArrowDropUp />
                     ))}
-                </th>
+                </th> */}
                 <th
                   style={{ fontWeight: "bold" }}
                   className="sortable-header"
@@ -383,19 +394,7 @@ const IndexTicket = () => {
                       <ArrowDropUp />
                     ))}
                 </th>
-                <th
-                  style={{ fontWeight: "bold" }}
-                  className="sortable-header"
-                  onClick={() => handleSortChange("createdAt")}
-                >
-                  Create Date
-                  {sortBy === "createdAt" &&
-                    (sortDirection === "asc" ? (
-                      <ArrowDropDown />
-                    ) : (
-                      <ArrowDropUp />
-                    ))}
-                </th>
+
                 <th
                   style={{ fontWeight: "bold" }}
                   className="sortable-header"
@@ -438,6 +437,19 @@ const IndexTicket = () => {
                 <th
                   style={{ fontWeight: "bold" }}
                   className="sortable-header"
+                  onClick={() => handleSortChange("createdAt")}
+                >
+                  Created Date
+                  {sortBy === "createdAt" &&
+                    (sortDirection === "asc" ? (
+                      <ArrowDropDown />
+                    ) : (
+                      <ArrowDropUp />
+                    ))}
+                </th>
+                <th
+                  style={{ fontWeight: "bold" }}
+                  className="sortable-header"
                   onClick={() => handleSortChange("dueTime")}
                 >
                   Due Date{" "}
@@ -448,10 +460,11 @@ const IndexTicket = () => {
                       <ArrowDropUp />
                     ))}
                 </th>
+                <th style={{ fontWeight: "bold" }}></th>
               </tr>
             </MDBTableHead>
             <MDBTableBody className="bg-light">
-              {dataTickets.map((ticket, index) => {
+              {dataTickets?.map((ticket, index) => {
                 const isSelected = selectedTickets.includes(ticket.id);
                 const ticketStatusOption = TicketStatusOptions.find(
                   (option) => option.id === ticket.ticketStatus
@@ -466,12 +479,8 @@ const IndexTicket = () => {
                         onChange={() => handleSelectTicket(ticket.id)}
                       />
                     </td>
-                    <td>
-                      <ViewCompact
-                        onClick={() => handleOpenEditTicket(ticket.id)}
-                      />
-                    </td>
-                    <td> {ticket.id}</td>
+
+                    {/* <td> {ticket.id}</td> */}
                     <td
                       className="tooltip-cell"
                       title={`Id:${ticket.id} \nDescription:${
@@ -485,12 +494,21 @@ const IndexTicket = () => {
                     <td>
                       {ticket.requester.lastName} {ticket.requester.firstName}
                     </td>
-                    <td>{formatDate(ticket.createdAt)}</td>
+
                     <td>
                       {ticket.type === "offline" ? (
-                        <Chip label="Offline" size="small" />
+                        <Chip
+                          label="Offline"
+                          size="small"
+                          sx={{ backgroundColor: "#c2c2c2" }}
+                        />
                       ) : (
-                        <Chip label="Online" size="small" variant="outlined" />
+                        <Chip
+                          label="Online"
+                          size="small"
+                          variant="outlined"
+                          sx={{ backgroundColor: "greenyellow" }}
+                        />
                       )}
                     </td>
                     <td>
@@ -540,7 +558,13 @@ const IndexTicket = () => {
                         </span>
                       }
                     </td>
+                    <td>{formatDate(ticket.createdAt)}</td>
                     <td>{formatDate(ticket.dueTime)}</td>
+                    <td>
+                      <ViewCompact
+                        onClick={() => handleOpenEditTicket(ticket.id)}
+                      />
+                    </td>
                   </tr>
                 );
               })}
